Add user models, with unit tests
Overridden Django default User model, to specify an email and password
field only.

Provides a simpler sign in, and also necessary to create a custom user model
before making first migrations, in the case that we might want to change
anything in future.Add user models, with unit tests

Overridden Django default User model, to specify an email and password
field only.

Provides a simpler sign in, and also necessary to create a custom user model
before making first migrations, in the case that we might want to change
anything in future.